评论模块初步完成

+++ model/comment.go
@@ -6,7 +6,7 @@ import "gorm.io/gorm"
 
 type Comment struct {
 	gorm.Model
-	Uid        int64  `json:"uid"`                   // 用户id
+	Uid        uint   `json:"uid"`                   // 用户id
 	Vid        int64  `json:"vid"`                   // 视频id
 	Content    string `json:"content,omitempty"`     // 评论内容
 	CreateDate string `json:"create_date,omitempty"` // 评论发布日期，格式 mm-dd
